feat: add securiti part and TDD

+++ Tax-Calculator/src/main/java/br/com/catalisa/zup/Tax/Calculator/Services/Tax/DeleteTax.java
@@ -8,10 +8,15 @@ import org.springframework.stereotype.Service;
 @AllArgsConstructor
 public class DeleteTax {
     private final TaxRepository taxRepository;
+
     public void deleteTax(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("Tax ID cannot be null");
+        }
         if (!taxRepository.existsById(id)) {
             throw new IllegalArgumentException("Tax not find");
         }
         taxRepository.deleteById(id);
     }
 }
+
